chore: rearrange react runtime code
