+++ .changeset/clever-books-attack.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+Reaaranged files inside the react runtime folder.
