+++ packages/runtime/src/runtimes/react/controls/number.ts
@@ -1,5 +1,10 @@
 import { match } from 'ts-pattern'
-import { NumberControlData, NumberControlDataTypeKey, NumberControlDataTypeValueV1, NumberControlDefinition } from '../../../controls'
+import {
+  NumberControlData,
+  NumberControlDataTypeKey,
+  NumberControlDataTypeValueV1,
+  NumberControlDefinition,
+} from '../../../controls'
 
 export type NumberControlValue<T extends NumberControlDefinition> =
   undefined extends T['config']['defaultValue'] ? number | undefined : number
@@ -8,9 +13,10 @@ export function useNumber<T extends NumberControlDefinition>(
   data: NumberControlData | undefined,
   definition: T,
 ): NumberControlValue<T> {
-  const value: number | undefined = match(data)
-    .with({ [NumberControlDataTypeKey]: NumberControlDataTypeValueV1 }, (val) => val.value)
-    .otherwise(val => val) ?? definition.config.defaultValue
+  const value: number | undefined =
+    match(data)
+      .with({ [NumberControlDataTypeKey]: NumberControlDataTypeValueV1 }, val => val.value)
+      .otherwise(val => val) ?? definition.config.defaultValue
 
   return value as NumberControlValue<T>
 }
