+++ packages/runtime/src/runtimes/react/controls/text-area.ts
@@ -1,5 +1,10 @@
 import { match } from 'ts-pattern'
-import { TextAreaControlData, TextAreaControlDataTypeKey, TextAreaControlDataTypeValueV1, TextAreaControlDefinition } from '../../../controls'
+import {
+  TextAreaControlData,
+  TextAreaControlDataTypeKey,
+  TextAreaControlDataTypeValueV1,
+  TextAreaControlDefinition,
+} from '../../../controls'
 
 export type TextAreaControlValue<T extends TextAreaControlDefinition> =
   undefined extends T['config']['defaultValue'] ? string | undefined : string
@@ -8,9 +13,10 @@ export function useTextAreaValue<T extends TextAreaControlDefinition>(
   data: TextAreaControlData | undefined,
   definition: T,
 ): TextAreaControlValue<T> {
-  const value: string | undefined = match(data)
-  .with({ [TextAreaControlDataTypeKey]: TextAreaControlDataTypeValueV1 }, (val) => val.value)
-  .otherwise(val => val) ?? definition.config.defaultValue
+  const value: string | undefined =
+    match(data)
+      .with({ [TextAreaControlDataTypeKey]: TextAreaControlDataTypeValueV1 }, val => val.value)
+      .otherwise(val => val) ?? definition.config.defaultValue
 
-return value as TextAreaControlValue<T>
+  return value as TextAreaControlValue<T>
 }
