+++ packages/runtime/src/runtimes/react/controls/text-input.ts
@@ -1,5 +1,10 @@
 import { match } from 'ts-pattern'
-import { TextInputControlData, TextInputControlDataTypeKey, TextInputControlDataTypeValueV1, TextInputControlDefinition } from '../../../controls'
+import {
+  TextInputControlData,
+  TextInputControlDataTypeKey,
+  TextInputControlDataTypeValueV1,
+  TextInputControlDefinition,
+} from '../../../controls'
 
 export type TextInputControlValue<T extends TextInputControlDefinition> =
   undefined extends T['config']['defaultValue'] ? string | undefined : string
@@ -8,9 +13,10 @@ export function useTextInputValue<T extends TextInputControlDefinition>(
   data: TextInputControlData | undefined,
   definition: T,
 ): TextInputControlValue<T> {
-  const value: string | undefined = match(data)
-    .with({ [TextInputControlDataTypeKey]: TextInputControlDataTypeValueV1 }, (val) => val.value)
-    .otherwise(val => val) ?? definition.config.defaultValue
+  const value: string | undefined =
+    match(data)
+      .with({ [TextInputControlDataTypeKey]: TextInputControlDataTypeValueV1 }, val => val.value)
+      .otherwise(val => val) ?? definition.config.defaultValue
 
   return value as TextInputControlValue<T>
 }
