+++ packages/runtime/src/runtimes/react/hooks/use-breakpoints.ts
@@ -1,0 +1,7 @@
+import { Breakpoints } from '../../../state/modules/breakpoints'
+import { getBreakpoints } from '../../../state/react-page'
+import { useSelector } from './use-selector'
+
+export function useBreakpoints(): Breakpoints {
+  return useSelector(state => getBreakpoints(state))
+}
