+++ packages/runtime/src/runtimes/react/hooks/use-component.ts
@@ -1,0 +1,6 @@
+import { ComponentType, getReactComponent } from '../../../state/react-page'
+import { useSelector } from './use-selector'
+
+export function useComponent(type: string): ComponentType | null {
+  return useSelector(state => getReactComponent(state, type))
+}
