+++ packages/runtime/src/runtimes/react/hooks/use-disable-register-element.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+
+export const DisableRegisterElement = createContext(false)
+
+export function useDisableRegisterElement() {
+  return useContext(DisableRegisterElement)
+}
