+++ packages/runtime/src/runtimes/react/hooks/use-document-cycles.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+
+export const DocumentCyclesContext = createContext<string[]>([])
+
+export function useDocumentCycles() {
+  return useContext(DocumentCyclesContext)
+}
