+++ packages/runtime/src/runtimes/react/hooks/use-document-key.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+
+export const DocumentContext = createContext<string | null>(null)
+
+export function useDocumentKey(): string | null {
+  return useContext(DocumentContext)
+}
