+++ packages/runtime/src/runtimes/react/hooks/use-document.ts
@@ -1,0 +1,6 @@
+import { Document, getDocument } from '../../../state/react-page'
+import { useSelector } from './use-selector'
+
+export function useDocument(documentKey: string): Document | null {
+  return useSelector(state => getDocument(state, documentKey))
+}
