+++ packages/runtime/src/runtimes/react/hooks/use-is-in-builder.ts
@@ -1,0 +1,6 @@
+import { getIsInBuilder } from '../../../state/react-page'
+import { useSelector } from './use-selector'
+
+export function useIsInBuilder(): boolean {
+  return useSelector(state => getIsInBuilder(state))
+}
