+++ packages/runtime/src/runtimes/react/hooks/use-is-preview.ts
@@ -1,0 +1,6 @@
+import { getIsPreview } from '../../../state/react-page'
+import { useSelector } from './use-selector'
+
+export function useIsPreview(): boolean {
+  return useSelector(state => getIsPreview(state))
+}
