+++ packages/runtime/src/runtimes/react/hooks/use-store.ts
@@ -1,0 +1,9 @@
+import { createContext, useContext } from 'react'
+import { ReactRuntime } from '../react-runtime'
+import { Store } from '../../../state/react-page'
+
+export const StoreContext = createContext(ReactRuntime.store)
+
+export function useStore(): Store {
+  return useContext(StoreContext)
+}
